Get next business Wednesday

Get next business Wednesday

Usage

calendar_nextBusinessWednesday(start_date = Sys.Date(), country_code = "PT")

Arguments

start_date

String (Default Sys.Date()). Reference date.

country_code

String (Default PT). Country code in the format ISO 3166-1 alpha-2. When provided, public holidays are considered.

Value

Date

Details

Find the next Wednesday that is not a holiday. When country is given, public holidays are considered, using Nager.Date API.

Examples

if (FALSE) { # \dontrun{
next_wednesday = GTFShift::calendar_nextBusinessWednesday(country_code="PT")
} # }