Get aggregated frequency per hour for each bus stop

For each stop, returns the number of departures aggregated per hour.

get_stop_hour_frequency(
  gtfs,
  date = NULL,
  route_types = list(0, 3, 5, 11),
  prepend_agency = TRUE
)

Arguments

gtfs

tidygtfs. GTFS loaded using tidytransit::read_gtfs.

date

Reference date to consider when analysing the GTFS file. Defaults to next business wednesday (in Portugal).

route_types

Restricts analysis to defined route_types, defaults to those that have conflicts on urban environments: tram and bus.

prepend_agency

Boolean. When true, stop_id is prepended with agency_id to avoid duplicate stop ids from multiple GTFS merging.

Value

A `data.frame` object with the following columns: - `stop_id`, the `stop_id` attribute from `stops.txt` file; - `hour`, the hour for which the frequency applies (24 hour format); - `frequency`, the number of services provided at the stop for the corresponding 60 minutes period; - `geometry`, the stop coordinates.

Details

This method analyses the GTFS feed for a representative day, generating for each stop the number of services aggregated per hour. For a detailed example, see the vignette vignette("get_stop_hour_frequency").

See also

Examples

if (FALSE) { # \dontrun{
gtfs <- tidytransit::read_gtfs("gtfs.zip")
frequency_analysis <- get_stop_hour_frequency(gtfs)
} # }