Read GTFS feed, fixing integrity errors

Read GTFS feed, fixing integrity errors

Usage

load_feed(path, store_path = NA)

Arguments

path

String. The location of the GTFS zip file. Either local or URL.

store_path.

String. (Optional) If provided, GTFS feed zip is stored at location. The file is overwritten if it already exists.

Value

A tidygtfs object.

Details

In addition to loading the GTFS feed, this method validates its integrity and applies the proper corrections if it does not comply with the following validations:

`stop_times.txt` with empty `arrival_time` or `departure_time`, filtering rows that do not comply.

Feeds with missing `shapes.txt` file, generating it using `GTFSwizard::get_shapes()`.

See also

Examples

if (FALSE) { # \dontrun{
gtfs <- GTFShift::load_feed("https://operator.com/gtfs.zip")
} # }