Aggregate lines based on overlap with simplified network

Aggregate lines based on overlap with simplified network

Usage

network_overline(
  network,
  df,
  attr,
  network_segment_length = 100,
  fun = sum,
  join_dist = 10
)

Arguments

network

sf. A spatial object representing the simplified network.

df

sf. A spatial object representing the lines to aggregate.

attr

String. The attribute to aggregate the lines by.

network_segment_length

Integer (Default 100). If not NA, network is split in segments of defined meters.

fun

Method (Default base::sum). Function to summarise the attributes by.

join_dist

Integer (Default 10). Meters to consider when joining routes and network segments.

Value

A spatial object of the network provided, extended with the aggregated values. A segment column will be also added to the network, as a unique identifier used for the aggregation process.

Details

This method allows for the lines aggregation. Given a simplified network, it identifies (using stplanr::rnet_join()) the segments corresponding to each line and uses them to aggregate the attribute defined in the parameters.

See also

Examples